ERA and FIER LOOP changes
ERA arrives earlier in the pipeline. It is faster to order data from ECMWF for the extent of the RTCs directly, rather than everytime you subset. The global datafile for the duration of Sentinel-1 mission is 333Gb so out of the question to download that.

NOTE: ERA5 download takes up to 2h, not because it's a big file but because it takes a while for ECMWF to quiery files in general, even very small ones.
When you subset, make sure it isn't too big: the REOF eats a LOT of RAM. If you combined dataset is 16Gb it's almost killing a 96Gb kernel.